# Improve `basic_string::find_first_of` and `basic_string::find_last_of` vectorization for large needles or very large haystacks