--- a/stl/src/vector_algorithms.cpp
+++ b/stl/src/vector_algorithms.cpp
@@ -3809,7 +3809,15 @@ namespace {
                 }
 
                 const int _Last_part_size_el = static_cast<int>(_Last_part_size / sizeof(_Ty));
-                const __m128i _Data1         = _mm_loadu_si128(reinterpret_cast<const __m128i*>(_Haystack));
+                __m128i _Data1;
+
+                if (_Haystack_length_bytes >= 16) {
+                    _Data1 = _mm_loadu_si128(reinterpret_cast<const __m128i*>(_Haystack));
+                } else {
+                    alignas(16) uint8_t _Tmp1[16];
+                    memcpy(_Tmp1, _Haystack, _Haystack_length_bytes);
+                    _Data1 = _mm_load_si128(reinterpret_cast<const __m128i*>(_Tmp1));
+                }
 
                 if (_mm_cmpestrc(_Data2, _Needle_length_el, _Data1, _Last_part_size_el, _Op)) {
                     return _mm_cmpestri(_Data2, _Needle_length_el, _Data1, _Last_part_size_el, _Op);
@@ -3866,7 +3874,16 @@ namespace {
                 }
 
                 const int _Last_part_size_el = static_cast<int>(_Last_part_size / sizeof(_Ty));
-                const __m128i _Data1         = _mm_loadu_si128(reinterpret_cast<const __m128i*>(_Haystack));
+                __m128i _Data1;
+
+                if (_Haystack_length_bytes >= 16) {
+                    _Data1 = _mm_loadu_si128(reinterpret_cast<const __m128i*>(_Haystack));
+                } else {
+                    alignas(16) uint8_t _Tmp1[16];
+                    memcpy(_Tmp1, _Haystack, _Haystack_length_bytes);
+                    _Data1 = _mm_load_si128(reinterpret_cast<const __m128i*>(_Tmp1));
+                }
+
                 _Test_whole_needle(_Data1, _Last_part_size_el);
 
                 return static_cast<size_t>(_Found_pos);