@StephanTLavavej 2024-10-20
> `__std_find_last_of_trivial_pos_N` has been shipped in https://github.com/microsoft/STL/pull/4934.

That's in VS 2022 17.13 Preview 2, which has not yet branched for release. That's happening soon (this Friday, 2024-10-25), but we're still going to have a Preview 3 open for at least a couple of weeks, so there's still time to mess with this function if it's important.