@AlexGuteniev 2024-10-20
> so there's still time to mess with this function if it's important

That's up to you which opportunity to use.

My understanding it is only a slight perf degradation at most, so I wouldn't want to rename it, as well as wouldn't merge a late change to a version shipped soon.