@AlexGuteniev 2024-10-21
A very cautious approach would be to keep `__std_find_last_of_trivial_pos_N` as it was introduced, and name all functions that try the bitmap somehow differently, like `__std_find_last_of_trivial_str_pos_N`  and `__std_find_first_of_trivial_str_pos_N`.