@StephanTLavavej 2024-11-09
I pushed a conflict-free merge with `main` and various small changes after a highly incomplete review (as I'm going to have to return to this after context-switching). I have one outstanding question about variable naming.