@AlexGuteniev 2024-11-09
> I have one outstanding question about variable naming.

Pushed significant changes to the function in question to avoid that variable, hopefully it is now clearer.