@AlexGuteniev 2024-11-09
No problem, I wasn't able to implement it in one round either, and I have a question to one of your questions!