@AlexGuteniev 2024-11-09
@StephanTLavavej mentioned on Discord that it is ok to use three-way strategy function.

As the implementation of it still needs to be established (there was one but it matched previous AVX bitmap implementation), moving to work in progress again.