@StephanTLavavej 2024-12-02
<details><summary>Click to expand 5950X benchmark results:</summary>

Benchmark                                                    | Before  |   After | Speedup
-------------------------------------------------------------|--------:|--------:|--------
`bm<AlgType::std_func, uint8_t>/2/3`                         | 5.74 ns | 6.08 ns | 0.94
`bm<AlgType::std_func, uint8_t>/6/81`                        |  117 ns |  115 ns | 1.02
`bm<AlgType::std_func, uint8_t>/7/4`                         | 11.1 ns | 12.6 ns | 0.88
`bm<AlgType::std_func, uint8_t>/9/3`                         | 13.8 ns | 13.8 ns | 1.00
`bm<AlgType::std_func, uint8_t>/22/5`                        | 14.4 ns | 13.8 ns | 1.04
`bm<AlgType::std_func, uint8_t>/58/2`                        | 15.7 ns | 15.4 ns | 1.02
`bm<AlgType::std_func, uint8_t>/75/85`                       | 35.2 ns | 35.9 ns | 0.98
`bm<AlgType::std_func, uint8_t>/102/4`                       | 16.6 ns | 15.8 ns | 1.05
`bm<AlgType::std_func, uint8_t>/200/46`                      | 44.5 ns | 44.6 ns | 1.00
`bm<AlgType::std_func, uint8_t>/325/1`                       | 8.67 ns | 10.4 ns | 0.83
`bm<AlgType::std_func, uint8_t>/400/50`                      |  111 ns |  114 ns | 0.97
`bm<AlgType::std_func, uint8_t>/1011/11`                     | 59.5 ns | 53.1 ns | 1.12
`bm<AlgType::std_func, uint8_t>/1280/46`                     |  241 ns |  251 ns | 0.96
`bm<AlgType::std_func, uint8_t>/1502/23`                     |  203 ns |  204 ns | 1.00
`bm<AlgType::std_func, uint8_t>/2203/54`                     |  559 ns |  525 ns | 1.06
`bm<AlgType::std_func, uint8_t>/3056/7`                      |  141 ns |  149 ns | 0.95
`bm<AlgType::std_func, uint16_t>/2/3`                        | 5.63 ns | 5.85 ns | 0.96
`bm<AlgType::std_func, uint16_t>/6/81`                       |  117 ns |  118 ns | 0.99
`bm<AlgType::std_func, uint16_t>/7/4`                        | 13.1 ns | 12.2 ns | 1.07
`bm<AlgType::std_func, uint16_t>/9/3`                        | 15.4 ns | 15.2 ns | 1.01
`bm<AlgType::std_func, uint16_t>/22/5`                       | 15.4 ns | 16.2 ns | 0.95
`bm<AlgType::std_func, uint16_t>/58/2`                       | 17.2 ns | 17.1 ns | 1.01
`bm<AlgType::std_func, uint16_t>/75/85`                      |  123 ns |  129 ns | 0.95
`bm<AlgType::std_func, uint16_t>/102/4`                      | 20.4 ns | 21.2 ns | 0.96
`bm<AlgType::std_func, uint16_t>/200/46`                     |  154 ns |  187 ns | 0.82
`bm<AlgType::std_func, uint16_t>/325/1`                      | 12.4 ns | 14.1 ns | 0.88
`bm<AlgType::std_func, uint16_t>/400/50`                     |  328 ns |  502 ns | 0.65
`bm<AlgType::std_func, uint16_t>/1011/11`                    |  257 ns |  300 ns | 0.86
`bm<AlgType::std_func, uint16_t>/1280/46`                    |  938 ns | 1120 ns | 0.84
`bm<AlgType::std_func, uint16_t>/1502/23`                    |  533 ns |  650 ns | 0.82
`bm<AlgType::std_func, uint16_t>/2203/54`                    | 1696 ns | 2178 ns | 0.78
`bm<AlgType::std_func, uint16_t>/3056/7`                     |  272 ns |  292 ns | 0.93
`bm<AlgType::std_func, uint32_t>/2/3`                        | 5.73 ns | 6.70 ns | 0.86
`bm<AlgType::std_func, uint32_t>/6/81`                       |  117 ns |  139 ns | 0.84
`bm<AlgType::std_func, uint32_t>/7/4`                        | 11.5 ns | 16.2 ns | 0.71
`bm<AlgType::std_func, uint32_t>/9/3`                        | 9.94 ns | 11.4 ns | 0.87
`bm<AlgType::std_func, uint32_t>/22/5`                       | 13.6 ns | 18.8 ns | 0.72
`bm<AlgType::std_func, uint32_t>/58/2`                       | 9.74 ns | 11.8 ns | 0.83
`bm<AlgType::std_func, uint32_t>/75/85`                      |  166 ns |  191 ns | 0.87
`bm<AlgType::std_func, uint32_t>/102/4`                      | 15.5 ns | 16.1 ns | 0.96
`bm<AlgType::std_func, uint32_t>/200/46`                     |  248 ns |  251 ns | 0.99
`bm<AlgType::std_func, uint32_t>/325/1`                      | 18.2 ns | 19.2 ns | 0.95
`bm<AlgType::std_func, uint32_t>/400/50`                     |  522 ns |  502 ns | 1.04
`bm<AlgType::std_func, uint32_t>/1011/11`                    |  282 ns |  271 ns | 1.04
`bm<AlgType::std_func, uint32_t>/1280/46`                    | 1459 ns | 1440 ns | 1.01
`bm<AlgType::std_func, uint32_t>/1502/23`                    |  846 ns |  823 ns | 1.03
`bm<AlgType::std_func, uint32_t>/2203/54`                    | 2967 ns | 2832 ns | 1.05
`bm<AlgType::std_func, uint32_t>/3056/7`                     |  340 ns |  359 ns | 0.95
`bm<AlgType::std_func, uint64_t>/2/3`                        | 6.22 ns | 5.72 ns | 1.09
`bm<AlgType::std_func, uint64_t>/6/81`                       |  121 ns |  122 ns | 0.99
`bm<AlgType::std_func, uint64_t>/7/4`                        | 13.0 ns | 11.8 ns | 1.10
`bm<AlgType::std_func, uint64_t>/9/3`                        | 9.82 ns | 9.82 ns | 1.00
`bm<AlgType::std_func, uint64_t>/22/5`                       | 15.4 ns | 12.4 ns | 1.24
`bm<AlgType::std_func, uint64_t>/58/2`                       | 18.7 ns | 12.4 ns | 1.51
`bm<AlgType::std_func, uint64_t>/75/85`                      |  352 ns |  362 ns | 0.97
`bm<AlgType::std_func, uint64_t>/102/4`                      | 29.0 ns | 32.0 ns | 0.91
`bm<AlgType::std_func, uint64_t>/200/46`                     |  500 ns |  524 ns | 0.95
`bm<AlgType::std_func, uint64_t>/325/1`                      | 44.8 ns | 50.6 ns | 0.89
`bm<AlgType::std_func, uint64_t>/400/50`                     | 1049 ns | 1075 ns | 0.98
`bm<AlgType::std_func, uint64_t>/1011/11`                    |  581 ns |  580 ns | 1.00
`bm<AlgType::std_func, uint64_t>/1280/46`                    | 3032 ns | 3120 ns | 0.97
`bm<AlgType::std_func, uint64_t>/1502/23`                    | 1790 ns | 1865 ns | 0.96
`bm<AlgType::std_func, uint64_t>/2203/54`                    | 6070 ns | 6541 ns | 0.93
`bm<AlgType::std_func, uint64_t>/3056/7`                     | 1029 ns | 1135 ns | 0.91
`bm<AlgType::str_member_first, char>/2/3`                    | 8.73 ns | 9.09 ns | 0.96
`bm<AlgType::str_member_first, char>/6/81`                   | 27.7 ns | 27.7 ns | 1.00
`bm<AlgType::str_member_first, char>/7/4`                    | 10.1 ns | 22.8 ns | 0.44
`bm<AlgType::str_member_first, char>/9/3`                    | 13.8 ns | 19.8 ns | 0.70
`bm<AlgType::str_member_first, char>/22/5`                   | 14.0 ns | 19.0 ns | 0.74
`bm<AlgType::str_member_first, char>/58/2`                   | 15.8 ns | 19.8 ns | 0.80
`bm<AlgType::str_member_first, char>/75/85`                  | 50.8 ns | 49.2 ns | 1.03
`bm<AlgType::str_member_first, char>/102/4`                  | 22.8 ns | 20.6 ns | 1.11
`bm<AlgType::str_member_first, char>/200/46`                 | 44.9 ns | 49.1 ns | 0.91
`bm<AlgType::str_member_first, char>/325/1`                  | 27.9 ns | 32.8 ns | 0.85
`bm<AlgType::str_member_first, char>/400/50`                 |  114 ns | 68.5 ns | 1.66
`bm<AlgType::str_member_first, char>/1011/11`                | 60.0 ns |  124 ns | 0.48
`bm<AlgType::str_member_first, char>/1280/46`                |  243 ns |  141 ns | 1.72
`bm<AlgType::str_member_first, char>/1502/23`                |  196 ns |  132 ns | 1.48
`bm<AlgType::str_member_first, char>/2203/54`                |  503 ns |  182 ns | 2.76
`bm<AlgType::str_member_first, char>/3056/7`                 |  141 ns |  222 ns | 0.64
`bm<AlgType::str_member_first, wchar_t>/2/3`                 | 11.8 ns | 12.7 ns | 0.93
`bm<AlgType::str_member_first, wchar_t>/6/81`                | 35.4 ns | 66.5 ns | 0.53
`bm<AlgType::str_member_first, wchar_t>/7/4`                 | 14.4 ns | 30.1 ns | 0.48
`bm<AlgType::str_member_first, wchar_t>/9/3`                 | 17.2 ns | 27.7 ns | 0.62
`bm<AlgType::str_member_first, wchar_t>/22/5`                | 17.9 ns | 27.9 ns | 0.64
`bm<AlgType::str_member_first, wchar_t>/58/2`                | 20.3 ns | 26.8 ns | 0.76
`bm<AlgType::str_member_first, wchar_t>/75/85`               | 70.6 ns | 56.0 ns | 1.26
`bm<AlgType::str_member_first, wchar_t>/102/4`               | 29.2 ns | 31.7 ns | 0.92
`bm<AlgType::str_member_first, wchar_t>/200/46`              |  156 ns | 64.4 ns | 2.42
`bm<AlgType::str_member_first, wchar_t>/325/1`               | 42.2 ns | 50.1 ns | 0.84
`bm<AlgType::str_member_first, wchar_t>/400/50`              |  252 ns | 74.7 ns | 3.37
`bm<AlgType::str_member_first, wchar_t>/1011/11`             |  261 ns |  128 ns | 2.04
`bm<AlgType::str_member_first, wchar_t>/1280/46`             |  590 ns |  166 ns | 3.55
`bm<AlgType::str_member_first, wchar_t>/1502/23`             |  675 ns |  194 ns | 3.48
`bm<AlgType::str_member_first, wchar_t>/2203/54`             |  969 ns |  242 ns | 4.00
`bm<AlgType::str_member_first, wchar_t>/3056/7`              |  264 ns |  305 ns | 0.87
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/2/3`      | 13.3 ns | 13.1 ns | 1.02
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/6/81`     |  125 ns | 32.7 ns | 3.82
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/7/4`      | 21.3 ns | 21.6 ns | 0.99
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/9/3`      | 17.5 ns | 22.2 ns | 0.79
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/22/5`     | 17.9 ns | 23.6 ns | 0.76
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/58/2`     | 19.6 ns | 25.2 ns | 0.78
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/75/85`    |  117 ns |  142 ns | 0.82
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/102/4`    | 22.8 ns | 28.7 ns | 0.79
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/200/46`   |  160 ns |  188 ns | 0.85
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/325/1`    | 42.1 ns | 55.8 ns | 0.75
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/400/50`   |  335 ns |  512 ns | 0.65
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/1011/11`  |  303 ns |  405 ns | 0.75
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/1280/46`  | 1062 ns | 1413 ns | 0.75
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/1502/23`  |  821 ns |  798 ns | 1.03
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/2203/54`  | 1979 ns | 2779 ns | 0.71
`bm<AlgType::str_member_first, wchar_t, L'\x03B1'>/3056/7`   |  269 ns |  362 ns | 0.74
`bm<AlgType::str_member_first, char32_t>/2/3`                | 11.9 ns | 14.9 ns | 0.80
`bm<AlgType::str_member_first, char32_t>/6/81`               | 35.7 ns | 38.2 ns | 0.93
`bm<AlgType::str_member_first, char32_t>/7/4`                | 13.5 ns | 24.0 ns | 0.56
`bm<AlgType::str_member_first, char32_t>/9/3`                | 13.0 ns | 23.9 ns | 0.54
`bm<AlgType::str_member_first, char32_t>/22/5`               | 14.8 ns | 22.1 ns | 0.67
`bm<AlgType::str_member_first, char32_t>/58/2`               | 12.5 ns | 22.9 ns | 0.55
`bm<AlgType::str_member_first, char32_t>/75/85`              | 71.2 ns | 50.4 ns | 1.41
`bm<AlgType::str_member_first, char32_t>/102/4`              | 18.5 ns | 28.4 ns | 0.65
`bm<AlgType::str_member_first, char32_t>/200/46`             |  112 ns | 54.7 ns | 2.05
`bm<AlgType::str_member_first, char32_t>/325/1`              | 24.1 ns | 42.1 ns | 0.57
`bm<AlgType::str_member_first, char32_t>/400/50`             |  198 ns | 75.9 ns | 2.61
`bm<AlgType::str_member_first, char32_t>/1011/11`            |  268 ns |  116 ns | 2.31
`bm<AlgType::str_member_first, char32_t>/1280/46`            |  566 ns |  141 ns | 4.01
`bm<AlgType::str_member_first, char32_t>/1502/23`            |  697 ns |  156 ns | 4.47
`bm<AlgType::str_member_first, char32_t>/2203/54`            |  960 ns |  203 ns | 4.73
`bm<AlgType::str_member_first, char32_t>/3056/7`             |  407 ns |  243 ns | 1.67
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/2/3`     | 14.9 ns | 11.7 ns | 1.27
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/6/81`    |  123 ns | 30.3 ns | 4.06
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/7/4`     | 19.8 ns | 18.2 ns | 1.09
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/9/3`     | 12.9 ns | 18.6 ns | 0.69
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/22/5`    | 15.0 ns | 21.5 ns | 0.70
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/58/2`    | 12.4 ns | 18.4 ns | 0.67
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/75/85`   |  174 ns |  181 ns | 0.96
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/102/4`   | 18.4 ns | 24.9 ns | 0.74
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/200/46`  |  247 ns |  272 ns | 0.91
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/325/1`   | 23.6 ns | 29.5 ns | 0.80
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/400/50`  |  489 ns |  495 ns | 0.99
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/1011/11` |  265 ns |  273 ns | 0.97
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/1280/46` | 1387 ns | 1409 ns | 0.98
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/1502/23` |  806 ns |  805 ns | 1.00
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/2203/54` | 2773 ns | 2885 ns | 0.96
`bm<AlgType::str_member_first, char32_t, U'\x03B1'>/3056/7`  |  341 ns |  352 ns | 0.97
`bm<AlgType::str_member_last, char>/2/3`                     | 8.54 ns | 7.87 ns | 1.09
`bm<AlgType::str_member_last, char>/6/81`                    | 26.5 ns | 25.4 ns | 1.04
`bm<AlgType::str_member_last, char>/7/4`                     | 9.84 ns | 21.1 ns | 0.47
`bm<AlgType::str_member_last, char>/9/3`                     | 13.9 ns | 19.2 ns | 0.72
`bm<AlgType::str_member_last, char>/22/5`                    | 14.6 ns | 19.1 ns | 0.76
`bm<AlgType::str_member_last, char>/58/2`                    | 16.0 ns | 19.4 ns | 0.82
`bm<AlgType::str_member_last, char>/75/85`                   | 46.3 ns | 42.8 ns | 1.08
`bm<AlgType::str_member_last, char>/102/4`                   | 16.3 ns | 20.0 ns | 0.82
`bm<AlgType::str_member_last, char>/200/46`                  | 44.9 ns | 42.2 ns | 1.06
`bm<AlgType::str_member_last, char>/325/1`                   | 34.3 ns | 30.4 ns | 1.13
`bm<AlgType::str_member_last, char>/400/50`                  |  156 ns | 54.4 ns | 2.87
`bm<AlgType::str_member_last, char>/1011/11`                 | 77.0 ns | 93.1 ns | 0.83
`bm<AlgType::str_member_last, char>/1280/46`                 |  237 ns |  119 ns | 1.99
`bm<AlgType::str_member_last, char>/1502/23`                 |  196 ns |  128 ns | 1.53
`bm<AlgType::str_member_last, char>/2203/54`                 |  497 ns |  177 ns | 2.81
`bm<AlgType::str_member_last, char>/3056/7`                  |  142 ns |  219 ns | 0.65
`bm<AlgType::str_member_last, wchar_t>/2/3`                  | 12.2 ns | 11.3 ns | 1.08
`bm<AlgType::str_member_last, wchar_t>/6/81`                 | 50.5 ns | 52.9 ns | 0.95
`bm<AlgType::str_member_last, wchar_t>/7/4`                  | 13.8 ns | 17.9 ns | 0.77
`bm<AlgType::str_member_last, wchar_t>/9/3`                  | 16.9 ns | 20.1 ns | 0.84
`bm<AlgType::str_member_last, wchar_t>/22/5`                 | 17.5 ns | 20.4 ns | 0.86
`bm<AlgType::str_member_last, wchar_t>/58/2`                 | 19.7 ns | 22.3 ns | 0.88
`bm<AlgType::str_member_last, wchar_t>/75/85`                | 75.9 ns | 49.1 ns | 1.55
`bm<AlgType::str_member_last, wchar_t>/102/4`                | 22.6 ns | 27.3 ns | 0.83
`bm<AlgType::str_member_last, wchar_t>/200/46`               |  120 ns | 55.3 ns | 2.17
`bm<AlgType::str_member_last, wchar_t>/325/1`                | 48.3 ns | 45.6 ns | 1.06
`bm<AlgType::str_member_last, wchar_t>/400/50`               |  207 ns | 68.9 ns | 3.00
`bm<AlgType::str_member_last, wchar_t>/1011/11`              |  446 ns |  125 ns | 3.57
`bm<AlgType::str_member_last, wchar_t>/1280/46`              |  576 ns |  153 ns | 3.76
`bm<AlgType::str_member_last, wchar_t>/1502/23`              |  724 ns |  171 ns | 4.23
`bm<AlgType::str_member_last, wchar_t>/2203/54`              |  993 ns |  224 ns | 4.43
`bm<AlgType::str_member_last, wchar_t>/3056/7`               |  266 ns |  276 ns | 0.96
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/2/3`       | 12.6 ns | 12.1 ns | 1.04
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/6/81`      |  109 ns | 27.2 ns | 4.01
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/7/4`       | 23.2 ns | 17.9 ns | 1.30
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/9/3`       | 17.0 ns | 20.0 ns | 0.85
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/22/5`      | 17.6 ns | 21.4 ns | 0.82
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/58/2`      | 20.2 ns | 24.1 ns | 0.84
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/75/85`     |  122 ns |  119 ns | 1.03
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/102/4`     | 23.5 ns | 26.2 ns | 0.90
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/200/46`    |  262 ns |  156 ns | 1.68
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/325/1`     | 77.5 ns | 51.6 ns | 1.50
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/400/50`    |  336 ns |  332 ns | 1.01
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/1011/11`   |  388 ns |  273 ns | 1.42
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/1280/46`   |  893 ns |  881 ns | 1.01
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/1502/23`   |  782 ns |  545 ns | 1.43
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/2203/54`   | 2100 ns | 1817 ns | 1.16
`bm<AlgType::str_member_last, wchar_t, L'\x03B1'>/3056/7`    |  267 ns |  275 ns | 0.97
</details>

I am indeed observing a nice speedup for the originally motivating `bm<AlgType::str_member_last, char>/400/50` case (2.87 speedup), but various regressions. `bm<AlgType::str_member_first, char>/1011/11` is a speedup of 0.48 (60.0 ns => 124 ns) and seems like a big haystack where we shouldn't be suffering from small-haystack effects.